Remove unnecessary Utils extraction
